# src/main/java/com/bux/investmentplans/domain/PlanStatus.java
package com.bux.investmentplans.domain;

/** See docs/detailed/1-architecture-overview.md 1.10 for the full lifecycle. */
public enum PlanStatus {
    ENTERED,
    ACTIVE,
    INACTIVE,
    CANCELED
}