# src/main/resources/openapi/investment-plans-api.yaml
openapi: 3.0.3
info:
  title: Investment Plans API
  description: >
    HTTP API for the Investment Plans Service
    (docs/detailed/1-architecture-overview.md 1.1, 1.3-1.4). Source of truth
    for the generated Spring controllers/models in this module — see the
    openapi-generator-maven-plugin execution in pom.xml.
  version: "0.1.0"

paths:
  /plans:
    post:
      operationId: createPlan
      tags: [ Plans ]
      summary: Create a plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePlanRequest'
      responses:
        '201':
          description: Plan created
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanCreatedResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
    get:
      operationId: listPlans
      tags: [ Plans ]
      summary: List the caller's plans
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PlanSummary'

  /plans/{planId}:
    parameters:
      - $ref: '#/components/parameters/PlanId'
    put:
      operationId: editPlan
      tags: [ Plans ]
      summary: Edit a plan (status-preserving; see PlanEdited, 1.10)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePlanRequest'
      responses:
        '200':
          description: Plan updated
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    get:
      operationId: getPlan
      tags: [ Plans ]
      summary: Get one plan
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanDetail'
        '404':
          $ref: '#/components/responses/NotFound'

  /plans/{planId}/executions:
    parameters:
      - $ref: '#/components/parameters/PlanId'
    get:
      operationId: listExecutions
      tags: [ Executions ]
      summary: List a plan's executions
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ExecutionSummary'
        '404':
          $ref: '#/components/responses/NotFound'

  /plans/{planId}/executions/{executionId}:
    parameters:
      - $ref: '#/components/parameters/PlanId'
      - name: executionId
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getExecution
      tags: [ Executions ]
      summary: Get one execution, including its orders (1.9)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionDetail'
        '404':
          $ref: '#/components/responses/NotFound'

components:
  parameters:
    PlanId:
      name: planId
      in: path
      required: true
      schema:
        type: string

  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: No such resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

  schemas:
    Error:
      type: object
      required: [ code, message ]
      properties:
        code:
          type: string
        message:
          type: string

    Investment:
      type: object
      description: One asset in a plan (1.9).
      required: [ assetId, quantity ]
      properties:
        assetId:
          type: string
        quantity:
          type: number

    CreatePlanRequest:
      type: object
      required: [ name, recurrenceDay, investments ]
      properties:
        name:
          type: string
        recurrenceDay:
          type: integer
          minimum: 1
          maximum: 28
        investments:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/Investment'

    PlanCreatedResponse:
      type: object
      required: [ planId ]
      properties:
        planId:
          type: string

    PlanStatus:
      type: string
      description: See docs/detailed/1-architecture-overview.md 1.10 for the full lifecycle.
      enum: [ ENTERED, ACTIVE, INACTIVE, CANCELED ]

    PlanSummary:
      type: object
      required: [ id, name, status, recurrenceDay ]
      properties:
        id:
          type: string
        name:
          type: string
        status:
          $ref: '#/components/schemas/PlanStatus'
        recurrenceDay:
          type: integer

    PlanDetail:
      allOf:
        - $ref: '#/components/schemas/PlanSummary'
        - type: object
          required: [ investments ]
          properties:
            investments:
              type: array
              items:
                $ref: '#/components/schemas/Investment'

    ExecutionStatus:
      type: string
      enum: [ INITIATED, PROCESSED ]

    ExecutionSummary:
      type: object
      required: [ id, planId, executionDay, status ]
      properties:
        id:
          type: string
        planId:
          type: string
        executionDay:
          type: string
          format: date
        status:
          $ref: '#/components/schemas/ExecutionStatus'
        filledCount:
          type: integer
        rejectedCount:
          type: integer

    OrderStatus:
      type: string
      enum: [ INITIATED, FILLED, REJECTED ]

    OrderSummary:
      type: object
      description: >
        (executionId, assetId) is this order's identity — no separate
        generated order id (2.9).
      required: [ assetId, quantity, status ]
      properties:
        assetId:
          type: string
        quantity:
          type: number
        status:
          $ref: '#/components/schemas/OrderStatus'

    ExecutionDetail:
      allOf:
        - $ref: '#/components/schemas/ExecutionSummary'
        - type: object
          required: [ orders ]
          properties:
            orders:
              type: array
              items:
                $ref: '#/components/schemas/OrderSummary'